requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-extensions==4.1

Pillow==12.0.0
requests==2.32.5

reportlab==4.4.6
xhtml2pdf==0.2.17
pypdf==6.6.0
